+++ lib/compiler/codegen.cpp
@@ -43,18 +43,18 @@ namespace amyinorbit::compass {
         writer.write<u32>(0xffffffff);
 
         // Heap
-        u32 heap_offset = writer.size();
+        u32 heap_offset = (u32)writer.size();
         writer.write<u16>(objects_.size());
         for(const Object* obj: objects_) {
             write_object(writer, obj);
         }
 
         // Globals
-        u32 globals_offset = writer.size();
+        u32 globals_offset = (u32)writer.size();
         writer.write<u16>(0);
 
         // Constant pool
-        u32 constants_offset = writer.size();
+        u32 constants_offset = (u32)writer.size();
         writer.write<u16>(constants_.size());
         for(const auto& val: constants_) {
             write_constant(writer, val);
